7/29/2017 12:54AM Push V0.0005
Wow it's been awhile.

Changes:

There are now 2 classes instead of just one to simplify coding the event
parsing. All event parsing related functions and statements are now in
their own class.

Modified the entire parsing system to be hopefully more effecient and
tidy.

Alot of information on the GUI has been removed due to the way I
restructured the Event parsing system. I need to make a new class to
handle the output of the Event Parser in order to properly display it.