7/29/2017 1:57PM Push V0.0006
Finished coding event handling from yesterday.

Created a new class to handle updating of variables that get displayed
to a GUI.

Created a new GUI that uses Java Swing instead of processing to draw
items to the screen, I should be able to ditch my dependency on
processing soon because of this. For now, once I make a proper
executable, don't close either window unless you want to close the
program, as if either one closes the program closes entirely.